rescale by expected volume

+++ nl/uva/imcn/algorithms/ConditionalShapeSegmentationFaster.java
@@ -3089,7 +3089,7 @@ public class ConditionalShapeSegmentationFaster {
                                 if (k>=18) if (score>0) score *= ISQRT3; else score /= ISQRT3;
                                 else if (k>=6) if (score>0) score *= ISQRT2; else score /= ISQRT2;
                                 //score = score/bestscore[obj];
-                                score = score/bestvol[obj];
+                                score = score/(float)bestvol[obj];
                                 heap.addValue(score,ngb,combinedLabels[best][idmap[ngb]]);
                                 best=nbest;
                             }
@@ -3134,7 +3134,7 @@ public class ConditionalShapeSegmentationFaster {
                                             // weight by relative size, so smaller structures are prioritized? not useful
                                             //newscore *= (1.0f-vol[obj]/bestvol[obj]);
                                             //newscore = newscore/bestscore[obj];
-                                            newscore = newscore/bestvol[obj];
+                                            newscore = newscore/(float)bestvol[obj];
                                             heap.addValue(newscore,ngb,combinedLabels[best][idmap[ngb]]);
                                             best=nbest;
                                         }
